update inc + src

+++ Core/Inc/encoder.h
@@ -3,6 +3,6 @@
 #include "stm32f411xe.h"
 
 void Encoder_Init(void);
-int16_t Encoder_GetDelta(void);
+int32_t Encoder_GetDelta(void);
 
 #endif
